+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support of synonyms.
 - Added ability to remove texts and translations.
 - Added ability to edit texts.
+- Extended edit page with search filter for translations.
 
 ## 1.0.0 - 2023-07-12
 - Initial version.